Add test for find availability, add test response for same.
Test for request to find availability for incident (SC Workforce management service)
has been added and a test response for the same has been provided.

The function is not available in the SDK, and details for implementing the same is also not
available yet. Till then, the test function will return the sample response.